windows/requirements-windows.txt
# Zusätzliche Abhängigkeiten nur für den Windows-Build (nicht für Docker/Linux)
pystray>=0.19.5
pillow>=10.0.0
pyinstaller>=6.0.0
